+++ app/api/search/route.ts
@@ -11,5 +11,6 @@ export async function GET(req: NextRequest) {
   const filePath = path.join(process.cwd(), `public/data`, `search_result_${page}.json`);
   const searchResult = readFileSync(filePath, 'utf-8');
 
+  return NextResponse.json({message: '서버 에러'}, {status: 500});
   return NextResponse.json(JSON.parse(searchResult));
 }
