+++ src/features/search-review/index.ts
@@ -1,1 +1,3 @@
-export {default} from './ui/SearchReviews';
+export {default as SearchReviews} from './ui/SearchReviews';
+export {default as SearchReviewsError} from './ui/SearchReviewsError';
+export {default as SearchReviewsLoading} from './ui/SearchReviewsLoading';
