+++ src/features/search-review/ui/SearchReviewsError.tsx
@@ -1,0 +1,7 @@
+export default function SearchReviewsError() {
+  return (
+    <section className="flex flex-col items-center p-4">
+      <h2>검색 결과를 받아오는데 실패했어요.</h2>
+    </section>
+  );
+}
